Fixes MOSIP-13227

+++ registration/registration-services/src/main/java/io/mosip/registration/dao/impl/RegPacketStatusDAOImpl.java
@@ -88,17 +88,20 @@ public class RegPacketStatusDAOImpl implements RegPacketStatusDAO {
 
 	@Override
 	public void delete(Registration registration) {
-		LOGGER.info("REGISTRATION - PACKET_STATUS_SYNC - REG_PACKET_STATUS_DAO", APPLICATION_NAME, APPLICATION_ID,
-				"Delete registration has been started");
+		LOGGER.info("Delete registration has been started");
 
 		AuditLogControl auditLogControl = auditLogControlDAO.get(registration.getId());
-
-		/* Delete Audit Logs */
-		auditLogControlDAO.delete(auditLogControl);
+		LOGGER.debug("Queried auditLogControl for registration {}, auditslogs: {}", registration.getId(), auditLogControl);
+		if(auditLogControl != null) {
+			/* Delete Audit Logs */
+			auditLogControlDAO.delete(auditLogControl);
+		}
 
 		/* Delete Registartion Transaction */
 		Iterable<RegistrationTransaction> iterableTransaction = registration.getRegistrationTransaction();
-		regTransactionRepository.deleteInBatch(iterableTransaction);
+		if(iterableTransaction != null) {
+			regTransactionRepository.deleteInBatch(iterableTransaction);
+		}
 
 		/* Delete Registartion */
 		registrationRepository.deleteById(registration.getId());
